moved ProcessorIdCache to a separate file.
